Windows - window functions

Windows - window functions

DSP.Windows.rect — Function.

rect(n)

Rectangular window function of length n.

source

DSP.Windows.hanning — Function.

hanning(n)

Hanning window of length n.

source

DSP.Windows.hamming — Function.

hamming(n)

Hamming window of length n.

source

DSP.Windows.tukey — Function.

tukey(n, alpha)

Tukey window of length n, parameterized by alpha. For alpha = 0, the window is equivalent to a rectangular window. For alpha = 1, the window is a Hann window.

source

DSP.Windows.cosine — Function.

cosine(n)

Cosine window of length n. Also called the sine window for obvious reasons.

source

DSP.Windows.lanczos — Function.

lanczos(n)

Lanczos window of length n.

source

DSP.Windows.triang — Function.

triang(n)

Triangular window of length n.

source

DSP.Windows.bartlett — Function.

bartlett(n)

Bartlett window of length n.

source

DSP.Windows.gaussian — Function.

gaussian(n, sigma)

Gaussian window of length n parameterized by the standard deviation sigma.

source

DSP.Windows.bartlett_hann — Function.

bartlett_hann(n)

Bartlett-Hann window of length n.

source

DSP.Windows.blackman — Function.

blackman(n)

"Exact" Blackman window, alpha = 0.16.

source

DSP.Windows.kaiser — Function.

kaiser(n, alpha)

Kaiser window of length n parameterized by alpha.

source

DSP.Windows.dpss — Function.

dpss(n, nw, ntapers=iceil(2*nw)-1)

The first ntapers discrete prolate spheroid sequences (Slepian tapers) as an n × ntapers matrix. The signs of the tapers follow the convention that the first element of the skew-symmetric (odd) tapers is positive. The time-bandwidth product is given by nw.

source

DSP.Windows.dpsseig — Function.

dpsseig(A, nw)

Eigenvalues of the DPSS matrix, representing the ratios of the power within the main lobe to the total power (main and sidelobes). A is the output of dpss, and nw is the time-bandwidth product provided to dpss as input.

source