Convolutions - similarity methods

Convolutions - similarity methods

DSP.conv — Function.

conv(u,v)

Convolution of two vectors. Uses FFT algorithm.

source

DSP.conv2 — Function.

conv2(u,v,A)

2-D convolution of the matrix A with the 2-D separable kernel generated by the vectors u and v. Uses 2-D FFT algorithm.

source

conv2(B,A)

2-D convolution of the matrix B with the matrix A. Uses 2-D FFT algorithm.

source

DSP.deconv — Function.

deconv(b,a) -> c

Construct vector c such that b = conv(a,c) + r. Equivalent to polynomial division.

source

DSP.xcorr — Function.

xcorr(u,v)

Compute the cross-correlation of two vectors.

source