LPC - Linear Predictive Coding

DSP.LPC.lpc — Function

lpc(x::AbstractVector, p::Int, [LPCBurg()])

Given input signal x and prediction order p, returns IIR coefficients a and average reconstruction error prediction_err. Note that this method does NOT return the leading 1 present in the true autocorrelative estimate; it omits it as it is implicit in every LPC estimate, and must be manually reintroduced if the returned vector should be treated as a polynomial.

The algorithm used is determined by the last optional parameter, and can be either LPCBurg or LPCLevinson.

DSP.LPC.lpc — Method

lpc(x::AbstractVector, p::Int, LPCBurg())

LPC (Linear-Predictive-Code) estimation, using the Burg method. This function implements the mathematics published in [1].

[1] - Enhanced Partial Tracking Using Linear Prediction (DAFX 2003 article, Lagrange et al) http://www.sylvain-marchand.info/Publications/dafx03.pdf

DSP.LPC.lpc — Method

lpc(x::AbstractVector, p::Int, LPCLevinson())

LPC (Linear-Predictive-Code) estimation, using the Levinson method. This function implements the mathematics described in [1].

[1] - The Wiener (RMS) Error Criterion in Filter Design and Prediction (N. Levinson, Studies in Applied Mathematics 25(1946), 261-278, https://doi.org/10.1002/sapm1946251261)